KW-659/refactor: did 생성 서비스 로직 디렉토리 수정/의존성 수정
